requirements.txt
langchain
langchain-community
langchain-groq
sqlalchemy
python-dotenv
streamlit
pandas
fastapi
uvicorn